-kulcs a szerver oldalon -kulcscsere agent és szerver között -agent le tudja kérni, és lementeni a szerver publikus kulcsát -agent tud regisztrálni a szerveren -agent ismereni a saját id-ját a szerveren -agent tud adatot feltölteni -szerver el tudja menteni az agentről jövő adatot db-be -mindez titkosítva -szerver a válaszban(ahol kell) objektumot küld xmlként -agent http post/get paramétereket küld a szervernek

+++ MonitoringAngent/src/diploma/vb/szt/agent/Memory.java
@@ -7,7 +7,7 @@ import org.hyperic.sigar.SigarException;
 public class Memory extends MonitoredItem
 {
 	private int sizeMb;
-	private float freePercent;
+	private int freePercent;
 
 	private final String SIZEMB = "sizeMb";
 	private final String FREEPERCENT = "freePercent";
@@ -17,7 +17,7 @@ public class Memory extends MonitoredItem
 		super("MEMORY");
 		Mem mem = sigar.getMem();
 		sizeMb = (int) (mem.getTotal() / 1024 / 1024);
-		freePercent = (float) mem.getFreePercent();
+		freePercent = (int) mem.getFreePercent();
 	}
 
 	@Override
